requirements.txt
Django==4.2.11
google-generativeai==0.3.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0 